ntp synchronization in MainActiviy.onCreate
